Create DetectionUpdatetoMetrics.js
this trigger is for converting any Detection to metrics & send it to recordstore.

so if a Detection happened it will be capture by this trigger and convert it to metrics base on the Detection Type, & afterwards, this trigger will convert those detection properties into metrics accordingly. 

The purpose of this trigger is to create a Dashboard based on any Detection type that has happened along the line from this trigger enabled to whenever you like it to be.  so we can create a dashboard that informs us on the specific timeline there are x detection category, x detection type, & so on.